Replace ::prepareSave with ::validateSave
This fixes newer MW versions and this is safe with MW 1.39.

This is deprecated in prepareSave and replaced with validateSave
so this should be safe.
