+++ src/MediaWiki/Content/SchemaContentHandler.php
@@ -3,12 +3,18 @@
 namespace SMW\MediaWiki\Content;
 
 use Content;
+use JsonContentHandler;
+use ParserOutput;
+use Status;
+use Title;
+use WikiPage;
+use MediaWiki\Content\ValidationParams;
 use MediaWiki\Content\Renderer\ContentParseParams;
 use MediaWiki\Content\Transform\PreSaveTransformParams;
-use JsonContentHandler;
+use MediaWiki\MediaWikiServices;
 use SMW\ParserData;
-use Title;
-use ParserOutput;
+use SMW\Schema\Exception\SchemaTypeNotFoundException;
+use SMW\Schema\Schema;
 
 /**
  * @license GNU GPL v2+
@@ -82,6 +88,91 @@ class SchemaContentHandler extends JsonContentHandler {
 			$pstParams->getUser(),
 			$pstParams->getParserOptions()
 		);
+	}
+
+	/**
+	 * @see ContentHandler::validateSave
+	 * @since 5.0
+	 *
+	 * {@inheritDoc}
+	 */
+	public function validateSave( Content $content, ValidationParams $validationParams ) {
+		$content->initServices();
+
+		$page = $validationParams->getPageIdentity();
+
+		if ( !( $page instanceof WikiPage ) ) {
+			$services = MediaWikiServices::getInstance();
+			$wikiPageFactory = $services->getWikiPageFactory();
+			$page = $wikiPageFactory->newFromTitle( $page );
+		}
+
+		$title = $page->getTitle();
+
+		$content->setTitlePrefix( $title );
+
+		$errors = [];
+		$schema = null;
+
+		try {
+			$schema = $content->getSchemaFactory()->newSchema(
+				$title->getDBKey(),
+				$content->toJson()
+			);
+		} catch ( SchemaTypeNotFoundException $e ) {
+			if ( !$content->isValid() && $content->getErrorMsg() !== '' ) {
+				$errors[] = [ 'smw-schema-error-json', $content->getErrorMsg() ];
+			} elseif ( $e->getType() === '' || $e->getType() === null ) {
+				$errors[] = [ 'smw-schema-error-type-missing' ];
+			} else {
+				$errors[] = [ 'smw-schema-error-type-unknown', $e->getType() ];
+			}
+		}
+
+		if ( $schema !== null ) {
+			$errors = $content->getSchemaFactory()->newSchemaValidator()->validate(
+				$schema
+			);
+
+			$schema_link = pathinfo(
+				$schema->info( Schema::SCHEMA_VALIDATION_FILE ) ?? '',
+				PATHINFO_FILENAME
+			);
+		}
+
+		$status = Status::newGood();
+
+		if ( $errors !== [] && $schema === null ) {
+			array_unshift( $errors, [ 'smw-schema-error-input' ] );
+		} elseif ( $errors !== [] ) {
+			array_unshift( $errors, [ 'smw-schema-error-input-schema', $schema_link ] );
+		}
+
+		foreach ( $errors as $error ) {
+
+			if ( isset( $error['property'] ) && $error['property'] === 'title_prefix' ) {
+
+				if ( isset( $error['enum'] ) ) {
+					$group = end( $error['enum'] );
+				} elseif ( isset( $error['const'] ) ) {
+					$group = $error['const'];
+				} else {
+					continue;
+				}
+
+				$error = [ 'smw-schema-error-title-prefix', "$group:" ];
+			}
+
+			if ( isset( $error['message'] ) ) {
+				$status->fatal( 'smw-schema-error-violation', $error['property'], $error['message'] );
+			} elseif ( is_string( $error ) ) {
+				$status->fatal( $error );
+			} else {
+				$status->fatal( ...$error );
+			}
+		}
+
+		return $status;
 	}
 
 	/**
