+++ tests/phpunit/MediaWiki/Content/SchemaContentTest.php
@@ -50,191 +50,12 @@ class SchemaContentTest extends \PHPUnit\Framework\TestCase {
 		);
 	}
 
-	public function testPrepareSave() {
-		$schema = $this->getMockBuilder( '\SMW\Schema\SchemaDefinition' )
-			->disableOriginalConstructor()
-			->getMock();
-
-		$schemaValidator = $this->getMockBuilder( '\SMW\Schema\SchemaValidator' )
-			->disableOriginalConstructor()
-			->getMock();
-
-		$schemaValidator->expects( $this->once() )
-			->method( 'validate' )
-			->will( $this->returnValue( [] ) );
-
-		$schemaFactory = $this->getMockBuilder( '\SMW\Schema\SchemaFactory' )
-			->disableOriginalConstructor()
-			->getMock();
-
-		$schemaFactory->expects( $this->any() )
-			->method( 'newSchema' )
-			->will( $this->returnValue( $schema ) );
-
-		$schemaFactory->expects( $this->any() )
-			->method( 'newSchemaValidator' )
-			->will( $this->returnValue( $schemaValidator ) );
-
-		$title = $this->getMockBuilder( '\Title' )
-			->disableOriginalConstructor()
-			->getMock();
-
-		$title->expects( $this->any() )
-			->method( 'getDBKey' )
-			->will( $this->returnValue( 'Foo' ) );
-
-		$page = $this->getMockBuilder( '\wikiPage' )
-			->disableOriginalConstructor()
-			->getMock();
-
-		$page->expects( $this->any() )
-			->method( 'getTitle' )
-			->will( $this->returnValue( $title ) );
-
-		$user = $this->getMockBuilder( '\User' )
-			->disableOriginalConstructor()
-			->getMock();
-
-		$parserOptions = $this->getMockBuilder( '\ParserOptions' )
-			->disableOriginalConstructor()
-			->getMock();
-
-		$instance = new SchemaContent(
-			json_encode( [ 'Foo' => 42 ] )
-		);
-
-		$instance->setServices( $schemaFactory );
-
-		$flags = '';
-		$parentRevId = '';
-
-		$this->assertInstanceof(
-			'\Status',
-			$instance->prepareSave( $page, $flags, $parentRevId, $user )
-		);
-	}
-
-	public function testPrepareSave_InvalidJSON() {
-		$schema = $this->getMockBuilder( '\SMW\Schema\SchemaDefinition' )
-			->disableOriginalConstructor()
-			->getMock();
-
-		$schemaValidator = $this->getMockBuilder( '\SMW\Schema\SchemaValidator' )
-			->disableOriginalConstructor()
-			->getMock();
-
-		$schemaValidator->expects( $this->once() )
-			->method( 'validate' )
-			->will( $this->returnValue( [] ) );
-
-		$schemaFactory = $this->getMockBuilder( '\SMW\Schema\SchemaFactory' )
-			->disableOriginalConstructor()
-			->getMock();
-
-		$schemaFactory->expects( $this->any() )
-			->method( 'newSchema' )
-			->will( $this->returnValue( $schema ) );
-
-		$schemaFactory->expects( $this->any() )
-			->method( 'newSchemaValidator' )
-			->will( $this->returnValue( $schemaValidator ) );
-
-		$title = $this->getMockBuilder( '\Title' )
-			->disableOriginalConstructor()
-			->getMock();
-
-		$title->expects( $this->any() )
-			->method( 'getDBKey' )
-			->will( $this->returnValue( 'Foo' ) );
-
-		$page = $this->getMockBuilder( '\wikiPage' )
-			->disableOriginalConstructor()
-			->getMock();
-
-		$page->expects( $this->any() )
-			->method( 'getTitle' )
-			->will( $this->returnValue( $title ) );
-
-		$user = $this->getMockBuilder( '\User' )
-			->disableOriginalConstructor()
-			->getMock();
-
-		$parserOptions = $this->getMockBuilder( '\ParserOptions' )
-			->disableOriginalConstructor()
-			->getMock();
-
-		$instance = new SchemaContent(
-			'Foo'
-		);
-
-		$instance->setServices( $schemaFactory );
-
-		$flags = '';
-		$parentRevId = '';
-
-		$this->assertInstanceof(
-			'\Status',
-			$instance->prepareSave( $page, $flags, $parentRevId, $user )
-		);
-	}
-
-	public function testSerializationOfClassInstance() {
-		$title = $this->getMockBuilder( '\Title' )
-			->disableOriginalConstructor()
-			->getMock();
-
-		$title->expects( $this->any() )
-			->method( 'getDBKey' )
-			->will( $this->returnValue( 'Foo' ) );
-
-		$page = $this->getMockBuilder( '\wikiPage' )
-			->disableOriginalConstructor()
-			->getMock();
-
-		$page->expects( $this->any() )
-			->method( 'getTitle' )
-			->will( $this->returnValue( $title ) );
-
-		$user = $this->getMockBuilder( '\User' )
-			->disableOriginalConstructor()
-			->getMock();
-
-		$parserOptions = $this->getMockBuilder( '\ParserOptions' )
-			->disableOriginalConstructor()
-			->getMock();
-
-		$instance = new SchemaContent(
-			json_encode( [ 'Foo' => 42 ] )
-		);
-
-		// Use an actual factory instance to ensure a "real" DB instance is
-		// invoked and would force a "RuntimeException: Database serialization
-		// may cause problems, since the connection is not restored on wakeup."
-		$instance->setServices( new \SMW\Schema\SchemaFactory() );
-
-		$flags = '';
-		$parentRevId = '';
-
-		$instance->prepareSave( $page, $flags, $parentRevId, $user );
-
-		$this->assertInternalType(
-			'string',
-			serialize( $instance )
-		);
-	}
-
 	public function testPreSaveTransform() {
-		$title = $this->getMockBuilder( '\Title' )
-			->disableOriginalConstructor()
-			->getMock();
+		$title = $this->createMock( '\Title' );
 
-		$user = $this->getMockBuilder( '\User' )
-			->disableOriginalConstructor()
-			->getMock();
+		$user = $this->createMock( '\User' );
 
-		$parserOptions = $this->getMockBuilder( '\ParserOptions' )
-			->disableOriginalConstructor()
-			->getMock();
+		$parserOptions = $this->createMock( '\ParserOptions' );
 
 		$instance = new SchemaContent(
 			json_encode( [ 'Foo' => 42 ] )
